Revert "Add action to cleanup older versions of the containers"
This reverts commit 038126074cfc0348e572152d45e3582d52f0efd6.

This was too complicate to make it work in a decent way: the publish step pushes
some metadata container, which floods the list of versions of the package, but
the actually useful one may get old if it isn't re-pushed in a while (maybe
because there haven't been updates to the TLOmodel repository), and so the
`actions/delete-package-versions` would actually delete the image we need.  I
couldn't find a way to obtain from the `docker/build-push-action` workflow the
sha we want to preserve.  Unless someone can figure out how to do preserve the
right image, I don't see how to make pruning work in a sensible way.